added a MUCH better system for scrolling up with little to no jitter :-).
If we have an analyzer, and the current address is known to be within a function then first we find the begining of that function.
Then, we attempt to disassemble from there until we run into the address we were on (stopping one instruction early).
this allows us to identify with good accuracy where the previous instruction was making upward scrolling more functional.

If all else fails, fall back on the old heuristic which works "ok"