Switching to C++11 always on. As of this commit, a C++11 capable compiler is necessary to compiler edb. This should enable more clear and concise code, with less dependancies on 3rd party libs.
Using new edisassm formatting API, this is more flexible and addresses some issues involving wanting to change
certain things at runtime instead of compile time

added an experiement "ByteArray" class, which will efficiently represent integer values of arbitrary sizes. The intent is
for this to enable a simple, consistant API for dealing with registers of different widths. Right now, it's a bit of
a mixed bag, and I'd like to simplify this going forward. This class is the first step