switching DebuggerCore::page_size() to return a constant. There are a couple of reasons for this:
1. while we could ask the system what the native page size is, why bother? It really
doesn't matter too much to the caller what the HARDWARE page size is. What matters is
that we have a consistant block size for which we can do bulk reads.

2. This debugger core plugin is VERY x86/x64 centric. We make system assumptions all
over the place in it, why should this be the exception. If we want to support some
exotic arch, we would simply make a new plugin for it.

3. Being a constant, means that in the future, it'll be easier to use for defining
buffer sizes and such. This goes along with my plans for a "Page Cache" architecture.

My plan for that is for read_bytes to just call read_pages which will use a cache.
Making rapid adjacent reads potentially much faster, I hope :-P.