Use execv instead of execvp.
Otherwise, when a file can't be executed (e.g. ELF64 on 32 bit kernel or
corrupted file) we get shell as a debuggee, which is very confusing.
The functionality we lose here was not used in EDB anyway, since it appears
to check existence of the file before calling execute_process(), so this
function never got any opportunity to search in $PATH.