Fix assembly of GAS from binutils 2.22
In binutils 2.22 GAS appears to output ELF file with .text section aligned to
4-byte boundary. This leads to extraneous NOPs before the instruction we are
trying to assemble, when %ADDRESS% is not aligned. This commit works around
this problem by employing a custom linker script with a SUBALIGN(0) command in
description of .text output section.

Another fix made here is to make sure the command returns error if as or ld
failed, instead of returning exit status of rm.