Fix comparison of regions to determine whether target is local
We use this function to avoid filename prefix when it's not needed. But the
same file may have multiple regions, e.g. .text and .data. This doesn't make
the reference from code to data non-local, since the prefix would be the same.
So it's more correct to compare region names.