Fix logic of detection of autorestarting system calls
I was wrongly assuming that the logic of restart is that (err&(EINTR|ERESTARTSYS))==true. Actually errno is never supposed to be used as a bitmask. Instead, there're four different ERESTART* error values, which correspond to the condition when the syscall is to be restarted. The EINTR|ERESTARTSYS is actually nothing other than ERESTART_RESTARTBLOCK, which is e.g. returned from nanosleep.