factored out "displacement_present", it was functionally equivalent to "disp != 0" since capstone adaptation.
made ArchProcessor not dependant on segment registers begin reported
as something suitable for an array index. This code is called rarely
(only when a segment override is being displayed), not much code depended on it
and it's really just moving the complexity from one spot to the other. So I figure
we'll make the typical case (no segment override) the one which has simple logic.