working on phasing out the "owner_" member of CapstoneEDB::Opcode". capstone operands don't have an explcit concept of being "owned" by an instruction, they are simply part of the struct via members.
Managing the owner makes implementing with proper copy/move constructors
difficult (I think they are currently, slightly bugged in some contexts)
so in the interest of thinning the wrapper and working towards less bugs,
I am going to phase owner_ out.

For the few cases (almost exclusively in ArchProcessor) where
we used operations that needed the owner, we'll pass it explicitly.

The vast majority of code remains uneffected.