Improve arrow-key selection navigation somewhat
This changeset makes it so that using the arrow keys to navigate within the current viewport does not need to call previous_instruction or forward_instruction.

It also makes it so that using the arrow keys to navigate downward past the viewport's bottom no longer causes the viewport view to shift down by a whole page; now we only scroll down 1/3rd of the viewport. This makes it less likely to get lost. This still isn't as great as I'd like it to be, but this will have to do for now since making it scroll the viewport down one instruction seems to have issues relating to instructions that were disassembled incorrectly.