Fixes issue #524 Implemented! (linux only for now)
`Debugger::next_debug_event()` now places an internal BP hook based on the debug information it can read from the ELF binary. Then
`Debugger::handle_trap()` tests if the BP it just hit was that internal BP.

For now, does some tests and then unconditionally tells the debugger to just continue. But in the future
we can allow the user to trap on load events, we can restore BPs in newly loaded libs, etc...

I do need to figure out if there is an efficient way to know WHICH library is being loaded.
We can scan the previous / current memory map and compare, but that seems inefficient.