Scaffolding for Analysis Persistence
This is work toward #478, but not a complete implementation yet.

In order to meaningfully serialize the analysis data of a particular region, we need to know:
* What file the region belongs to,
* What the base address is of that file in memory,
* And how that compares to where the file was actually loaded (relocation).

So first we needed to add `base_address()` public method to all of the `IBinaryInfo` plugins to satisfy the second requirement.  Then I added a method that generates a filepath that can be used to store the region's analysis. This is a little rough; we have to iterate all regions in memory to locate regions that belong to the same file as our analyzed region (because of the possibility of multiple code sections) and then we try to use BinaryInfo plugins on every region until one finally fits. We use that to obtain the base address and the load address of the module. Then, finally, we calculate a relative offset and put it into the filename.

Once this is in place we just need to actually implement serialization of the analysis data. This is where I got a little bamboozled, so I'm putting up what I have so far since I think what I have so far is pretty uncontroversal.