adding (but not used yet) a concept of a "RegisterRef".
This will make it MUCH easier to handle registers of various widths in a non-arch specific way.

The idea is simple, the State objects will return a RegisterRef, which is basically
an offset, size, mask tuple. The offset is will basically be something like:

offsetof(State, eax) // for the eax regiser of course

Then when arch specific code like in the ArchProcessor or RegiserView wants concrete values
it can combine this with an actual State object to extract the actual value. This way, we
don't need to have hacks to deal with things like "ARM doesn't have an 80-bit float type".

This is based on Mozilla's approach, which seems highly effective.