Add sume support for Thumb
This does only the most obvious things: re-initializes Capstone according to
current T flag. But in practice, we want to know which functions (blocks of text
segment) are Thumb and which are ARM, and disassemble them accordingly, unless
PC is is there, and T contradicts this heuristic.
GDB somehow does determine that e.g. main() is thumb, while _start isn't, I
guess this info is hidden in the ELF data structures.