Detect and report x32 system calls
This still doesn't make obvious the difference between e.g.
rt_sigreturn=15 (64-bit) for which the app might for some strange reason
set __X32_SYSCALL_BIT vs. rt_sigreturn=512 (x32), which is supposed to
be called with this bit set. Maybe the format of the analysis string
should be another.