introducing a beter way for things to capture debug events. instead of setting a pointer, possibly needing to call the old handler and then restoring the pointer when it's done, we simply add the handler to the top of a stack. If a handler doesn't want to deal with an event, it just returns DEBUG_NEXT_HANDLER and the system will continue calling the handlers in the chain until one handles the event.
This is safer, and more robust.

It is also step 1 in creating a better way to deal with events in general.
What I'd love to eventually get to is something that allows code like this:

edb::v1::resume(status_code, [](std::shared_ptr<IDebugEvent> &e) {
	// handle the next event directly or return DEBUG_NEXT_HANDLER
});

this would make a few places in the code MUCH easier to understand and maintain