introduce DebugEventHandlers (fix #650)
This commit fixes the issue #650:
Erasing an element of std::deque while iterating over it

The problem was that step out caused access to an invalid
iterator (a iterator to a erased element of g_DebugEventHandlers).
Here is what was going on:
1. execute_debug_event_handlers() is run.
2. execute_debug_event_handlers() is iterating over handlers
   and now stays on element corresponding RunUntilRet.
3. RunUntilRet::handle_event() is run, prints
   "Found ret; passing back to debugger" and calls "delete this".
4. RunUntilRet::~RunUntilRet() is called doing
   remove_debug_event_handler(this).
5. remove_debug_event_handler() erases the element corresponding
   to RunUntilRet. Erasing the element invalidates the iterator
   that is hold by execute_debug_event_handlers (!!!!!)
6. remove_debug_event_handler() quits, RunUntilRet::~RunUntilRet()
   quits, RunUntilRet::handle_event() quits.
7. The execution is back to execute_debug_event_handlers(). The next
   operation it does is ++it incrementing an invalid iterator.

This problem is solved by introducing a class called
DebugEventHandlers. It holds IDebugEventHandler* and allows removing
when execute_debug_event_handlers is being in progress.

Here is the copy of its description:

This class is a container of IDebugEventHandler*. It supports three
operations:
1. add -- Adding IDebugEventHandler*
2. remove -- Removing IDebugEventHandler*
3. execute -- Iterating over added handlers and calling them

The reason why a simple standard container is not enough is that
handlers can be added/removed while the list is being iterated on.
Naive implementation would cause accessing an invalid iterator in this
case.

This class supports adding new handlers during iteration. Removing
during iteration is also supported, but is restricted: only the
handler that is currectly being executed can be removed.

In order to support removing handlers during iteration the iterator
is incremented before calling corresponding handler.

Recursive calls to execute are forbidden.

Contract violations of this class are unrecoverable and cause
program termination (std::abort is called).