moved some concepts out of BinaryInfo and into PlatformProcess yes, it assumes ELF on linux for now, but it's WAY more reliable because now the code is using the AUX vector to properly calculate things like the load address instead of guessing.
I also moved things like "calulcate_main" to PlatformProcess as this
isn't REALLY a per module concept, we want the main that is in the
primary executable, and once again, basing this on infor we get from AUXV
is much more reliable and simpler.

Moved libELF and libPE to the main folder since they are just
"generally useful" for the debugger.