stubbing out detection of process_vm_readv/process_vm_writev viability.
Still need to sort out permissions, the API may or may not be usable for our needs
since I know that at least the write version fails based on page protections
even with ptrace permissions.

But either way, good to at least have the code to detect it in our pocket, the API
has been around for several years now.