So...
Long ago, we moved to using forward declarations and dynamic allocation for Ui::whatever
in an attempt to reduce compile times.

However, after looking at this carefully, I have concluded that for **plugins**
this is actually a pessimization! (both at runtime and compile time):

1. Plugins are fairly isolated architecturally, the main code NEVER directly includes
   their headers, so these files are included maybe twice. So the avoidance of repeated
   code processing is minimal at best.

2. The forward declares in the Ui namespace are ugly :-P

3. It requires dynamic allocation which is slower at runtime! Qt is already a pretty "new"
   happy framework, but for each plugin dialog we effectively have to do two news when we COULD
   be doing one. Dynamic allocations are SLOW!

4. Doing this using values requires less code. Since no dynamic allocation is needed, we get
   to have simpler constructors and defaulted destructors meaning the compiler actually has
   less code to process in a bunch of spots. It's small, but we've got a lot of code, it adds up

5. My compile time actually went down slightly (a couple of seconds, so marginal) using the value
   based includes.

So for **plugins**, we'll switch to the value based ui method. I think for the edb main executable
it might still be a net win, but I'm not sure.