fix a bug that symbols aren't loaded when executable path contains whitespace
When program with whitespace in its path is run, edb prints
the warning:

WARNING: File "/home/ivan/.cache/codef00.com/edb/symbols//home/ivan/My Documents/hello.map" seems corrupt

and the list of symbols is empty.

The bug is caused by the fact that the file path is read with
istream::operator>> that stops reading at the first whitespace. This
commit uses std::getline to read the line fully.