Implement fix for issues #627, #686, #706, and #742
Essentially, allow the user to specify where process input and output come from and go to.

The syntax is not as "user friendly" as it could be, but it works well.

Basically, we can now do this:

mkfifo my_pipe
./edb --stdin my_pipe --run /bin/cat

and then any data written to my_pip will be recieved by the cat process.

This allows edb to debug processes which need to have their input recieved from
something that generates the stream.

In the future, we COULD make this more advanced and suport things like:

./edb --run ./A | ./B

Where edb would attach to the B process, but would run A, having set up the
pipes to connect A's output to B's input... but that's a more complex story.